refactor(ciot): round road distance to whole kilometers
The distance suggestion and its cache now use whole km (unsigned
integer column) instead of one decimal place, matching how operators
report route distance.

+++ app/Models/CityDistance.php
@@ -17,7 +17,7 @@ class CityDistance extends Model
     protected function casts(): array
     {
         return [
-            'km' => 'float',
+            'km' => 'integer',
             'fetched_at' => 'datetime',
         ];
     }
